Simplify hosted review and admin surfaces
